+++ buildSrc/src/main/kotlin/dependency/network/Ktor.kt
@@ -6,5 +6,5 @@ object Ktor : Dependency {
 
     override val group = "io.ktor"
     override val artifact = "ktor"
-    override val version = "1.5.1"
+    override val version = "1.5.2"
 }